SocialServices-Refactoring: Moved dto package into model one.
