--- src/main/java/com/umssonline/social/common/dto/CreateRateDto.java
@@ -1,4 +1,0 @@
-package com.umssonline.social.common.dto;
-
-public class CreateRateDto {
-}
