+++ src/main/java/com/umssonline/social/controllers/CommentsRestController.java
@@ -1,7 +1,7 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.common.dto.CreateCommentDto;
-import com.umssonline.social.common.dto.UpdateCommentDto;
+import com.umssonline.social.models.dto.CreateCommentDto;
+import com.umssonline.social.models.dto.UpdateCommentDto;
 import com.umssonline.social.models.Comment;
 import org.springframework.http.ResponseEntity;
