+++ src/main/java/com/umssonline/social/controllers/MessagesRestControllerImpl.java
@@ -1,7 +1,7 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.common.dto.CreateMessageDto;
-import com.umssonline.social.common.dto.UpdateMessageDto;
+import com.umssonline.social.models.dto.CreateMessageDto;
+import com.umssonline.social.models.dto.UpdateMessageDto;
 import com.umssonline.social.models.Message;
 import com.umssonline.social.services.SocialService;
 import org.modelmapper.ModelMapper;
