+++ src/main/java/com/umssonline/social/controllers/RatesRestControllerImpl.java
@@ -1,7 +1,7 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.common.dto.CreateRateDto;
-import com.umssonline.social.common.dto.UpdateRateDto;
+import com.umssonline.social.models.dto.CreateRateDto;
+import com.umssonline.social.models.dto.UpdateRateDto;
 import com.umssonline.social.models.Rate;
 import com.umssonline.social.services.SocialService;
 import org.modelmapper.ModelMapper;
